[Feature] 게시글 생성 및 조회 API 연동

+++ feature/mypage/src/main/java/com/example/mypage/navigation/MyPageNavigation.kt
@@ -25,7 +25,7 @@ fun NavController.navigateToSetting(navOptions: NavOptions? = null) {
 
 fun NavGraphBuilder.myPageNavGraph(
     navigateToLogin : () -> Unit,
-    navigateToPost: () -> Unit,
+    navigateToPost: (Int) -> Unit,
     navigateToUpdateProfile: () -> Unit,
     navigateToSetting: () -> Unit,
     navigateBack: () -> Unit
